Finish lab2 the bomb_lab
This is a solution for the ECNU Dase's random bomb.You can find the solution and it`s explanation in solution.md or bomb_solve.txt .The bomb is in bomb.c but it`s a binary file.The dumped file is in dumped.txt .As for the original CMU bomb,it`s in the example folder. You can find it`s solution in the folder or you may check it on the internet.